requirements.txt
tensorflow
keras
numpy
pandas
matplotlib
torch
torchvision
transformers
streamlit
gradio
pillow
timm